Bug 1298756 - Use uint8_t/uint16_t as underlying type for enums from webidl. r=bz
MozReview-Commit-ID: 3uS9s5ZrPFd

UltraBlame original commit: a42a1279c2162fa2a6a0f4b448ede7f9d6b2cc6e

+++ layout/style/nsStyleConsts.h
@@ -314,8 +314,8 @@ enum class StyleOrient : uint8_t {
 
 
 namespace dom {
-enum class PlaybackDirection : uint32_t;
-enum class FillMode : uint32_t;
+enum class PlaybackDirection : uint8_t;
+enum class FillMode : uint8_t;
 }
